requirements.txt
boto3
requests
python-dotenv
beautifulsoup4
pandas
